requirements.txt
# ====================================================================
# LLM Choice in Graph-Based RAG Construction — pinned dependencies
# ====================================================================
# Generated against Python 3.12 on 2026-04-15.
# Reproduce the venv with:
#     python3 -m venv .venv
#     source .venv/bin/activate
#     pip install -U pip
#     pip install -r requirements.txt

# ---- LLM clients ----
openai==2.31.0
anthropic==0.95.0

# ---- Corpus + chunking ----
wikipedia==1.4.0
tiktoken==0.12.0

# ---- Data + graphs ----
pandas==2.3.3
numpy==2.4.4
networkx==3.6.1
scikit-learn==1.8.0

# ---- Plotting ----
matplotlib==3.10.8

# ---- Config / I/O ----
python-dotenv==1.2.2

# ---- Transitive (pinned to avoid resolver surprises) ----
httpx==0.28.1
pydantic==2.13.0
requests==2.33.1
urllib3==2.6.3
certifi==2026.2.25
